Add Google Provider
This commit re-adds the Google provider. Instead of the iOS and Android
configuration we are now using a Web Client for the authentication flow
within kubenav. This means that a new Web Client must be created in the
Google Cloud Console with "https://kubenav.io/google.html". The
generated client id and secret can then be used within kubenav to access
all Kubernetes clusters from all Google projects.

Note: The redirect url will be changed once the v4 version is official
published and the rework of the website is done.