Add OIDC Provider
This commit re-adds the OIDC provider. FOr the OIDC provider a user has
to provide a name, discovery url (issuer url) and a client id. If no
PKCE method is selected the user has to provide a client secret.

If the user provides a refresh token we will use it to generate a id
token to access the Kubernetes API for the configured cluster. If no
refresh token is provided the user has to click the sign in button to
generate a code, which can be used to get a refresh token.

Note: The current redirect url "https://kubenav.io/oidcv4.html" will
be changed before the final release of kubenav v4.