Add authentication
It is now possible to protect your configuration in kubenav via
authentication. Authentication can be enabled in the settings. When a
user tries to enable authentication we are checking if the device is
supported and enrolled in any authentication methods. If this is the
case authentication will be enabled, if not the user will see an error
message.

When authentication is enabled we use the new login page as initial
route instead of the home page. The login page just displayes a blank
screen where the user must authenticate. We try to authenticate a user
as long as the authentication succeded. Then we redirect the user to the
home page.