Fix json decoding errors
We had some parts in the app where the json decoding failed, because the
returned result from our API calls was null. We are now checking for
null values before we try to decode these values and for some calls also
after we decoded these values.

We also improved the logging, so that all the results from API calls and
calls of the Go code are printed, before we try to decode the result.
This should help us to debug potential problems.