Improve live metrics view
Instead of displaying the live metrics for CPU and Memory among
themselves, we are now using tabs. This means the view contains two tabs
one for the CPU values and another one for the Memory values and a user
can switch between these two graphs.

We also changed the style of the tab bar in the terminals widget to
match the new tab styling we are using in the live metrics view.