Show snackbar on user action
For operations which are taking a bit longer (retrieving logs, getting a
terminal and establishing a port forwarding session), we are now showing
a snackbar after the user initialized the action. This should help users
to see that sth. is running in the background an should avoid double
clicks on the corresponding action.