Show requests and limits in live metrics
The live metrics view also contains the requests and limits for the
selected Pod now. For the requests we display a yellow line and for the
limits a red line.

It is also possible to select a container and just display the
containers usage, requests and limits in the live metrics.

Closes #407