Add list to maintain favorite namespaces
It is now possible to maintain a list of favorite namespaces, within the
settings of the app. These namespaces will always be displayed on the
top of the namespace selection bottom sheet for quicker access.

They can also be used by users, which do not have the permissions to
list all namespaces in a cluster, so that they can maintain a list of
namespace where they have access to.