Fix form validation and view all Pods on Node
This commit fixes the form validation for several widgets. The problem
was that we called "finish(context)" also when the form validation
failed. Now we are validating the form, before we call
"finish(context)", to keep the BottomSheet for a form open.

We also fixed a small bug, when a user clicked on the view all link in
the node details to view all Pods. Until now this only worked when the
user selected all namespaces before. Now we are setting the namespace
parameter to an empty value to overwrite the users selection of the
namespace.