Add Prometheus Plugin
This commit re-adds the Prometheus plugin. In this first version it is
possible to configure a Prometheus instance and view some default
metrics for resources (e.g. cpu and memory usage of pods or the number
of pods for deployments).

It is also possible to define custom metrics for resources by adding a
"kubenav.io/prometheus" annotation to a manifest with a list of charts.