Rework State Management (#425)
Until now we used GetX for the state management in the app. Since we are
not very happy with this, we decided to remove GetX in all places and
use Provider for the state management. We also switched to the Navigator
instead of using GetX for the routing in the app. Last but not least we
also removed GetStorage and save the clusters, providers and app
settings via the Flutter Secure Storage package.

This commit also updates the required Dart and Go dependencies.

Note: The dark mode is not working in this version, because there will
be a larger rework of the theming within the next PR.