Rework Storage
We are now using the Flutter Secure Storage package only on iOS and
Android. On macOS, Linux and Windows we are using now the shared
preferences package to store the users settings and bookmarks.

To only use the Flutter Secure Storage package on iOS and Android we
are using a fork of the package where we removed the other
implementations: https://github.com/kubenav/flutter_secure_storage/tree/kubenav

We decided to not use the package for desktop, because it was not
working correctly for macOS, because the required provisioning profile
for the Keychain Sharing capability could not be added in the GitHub
Action :-(.

If someone nows how we can build the app with the correct provisioning
profile via our GitHub Action we can switch back to the Flutter Secure
Storage package on desktop.