Improve Pod Status (#431)
Until now a Pod was always shown as healthy when the status was
"Running". This is now changed, so that we also check the number of
containers that are ready and the number of container which should be
ready. If the numbers do not match the Pod is marked as unhealthy.