Improve Theme Handling (#433)
We are now using a ThemeRepository to handle the user selected theme in
the app and to apply the corresponding theme. A user can select between
a light and a dark theme. The dark theme was completly rework compared
to the former version, to hopefully look a bit nicer now.

We also adjusted the modal bottom sheet handling and snackbar handling,
so that the snackbar is now rendered above the modal bottom sheet.