Improve Cursor Style and Other Fixes
On some items the cursor style wasn't adjusted on dekstop, so that there
was no indication that the item is clickable. We also had the problem
that the clickable area was to large for some items (e.g app actions).
These issues should be fixed now.

We also removed the dismiss option from the snackbar, because sometimes
this could break the app, when the context isn't available anymore.

In some places we still had used "Theme.of", which is now replaced with
our new theme() function.