Use SVG instead of PNG Images
We are now using svg instead of png images for everything, except the
screenshots in the help section. This has the advantage that we can
reduce the size of the bundled assets and that we do not have to
maintain the images in multiple resolutions.

We also moved the "assets/_images" folder to "utils/images" to not mix
up the assets for the app with the files we need to create the assets.