Add Path Support for Prometheus
Until now it was not possible to specify the path of a Prometheus
instance, when the in-cluster configuration was used. This caused some
problems with Prometheus compatible APIs like VictoriaMetrics, where the
Prometheus API is exposed under a prefix (e.g. "/select/0/prometheus").

With the new path option, it is now possible to use such APIs. By
default the path is set to an empty string, which is the same behaviour
like before without the path option.