Add Show Clusters on Start Option (#437)
The new show clusters on start option can be used to show a list of
clusters when the app is started. By default we will directly show the
overview page when the app is started with the active cluster from the
users last session. When the show clusters on start option is enabled we
not show the overview page and instead show a list of all configured
clusters, where a user can select the active cluster before he can
continue.

The list of clusters is only shown when the corresponding option is
enabled and the list of clusters is not empty.