Fix AppNoClustersWidget Theming
When the user hasn't added a cluster yet and the AppNoClustersWidget was
shown in the setting and the user changed the theme the widget wasn't
adjusted. This is now fix by listining to theme changes in the widget.

We also adjusted the Home widget, to show an empty container when the
app is not initialized (isAuthenticated==false).