Improve Bookmarks Handling (#441)
Until now bookmarks for a cluser were not deleted, when the
corresponding cluster was deleted, which broke the app when a user tried
to open such a bookmark. This is now fixed by deleting all bookmarks
with the cluster id of the deleted cluster.

We now show the name of the cluster instead of the cluster id in the
bookmarks preview and the bookmarks list.