Fix Cluster Names for Google and AWS Provider (#444)
The cluster name shown in the list where a user could select the
clusters he wants to add and the name of the then added clusters were
not the same. This is now fixed by using the same naming schema in both
places.

This commit also improves the presentation of clusters in the select
clusters view with very long names, so that they are cutted of at any
character now, to show as much as possible from the name.