Improve Reorder Handling
This commit improves the reorder handling of clusters, namespaces and
bookmarks. Until now it was difficult to scroll in the clusters,
namespaces or bookmarks list, because this automatically triggered the
reordering. This is now fixed by adding dedicated drag handles to these
lists, which can be used to reorder them.

Note: The drag handle icon in the clusters list is the status indicator
icon. We decided to reuse this icon, so that we do not have to add two
different icons to the list.