Improve Selection of Clusters (#446)
When a cluster is added via the Google, AWS or Azure provider it is now
easier to select the clusters which should be added. Before this change
it was only possible to select the clusters by clicking on the
corresponding checkbox. Now it is also possible to select the cluster by
clicking on the name. This is possible because we are using the
"CheckboxListTile" now.