Improve Image in Help (#448)
We resized the images in the help section of the app to better fit the
screen. We also added some missing images for the Google Provider and
Namespace section.

Last but not least this commit also adds the documentation for the Helm
and Prometheus plugin. While writing the documentation we also found a
bug in the parsing of the yaml for the custom Prometheus charts and
dashboards which is also fixed now.