Fix Theming of Icons
Some icons were still using a static color instead of a color from our
theme. The color of the icons which were using "Colors.grey[300]" are
now using the primary or secondary text color of the selected theme.

The icons are using the primary text color when the text uses the
primary color. If the text uses the primary and secondary color the
icons will use the secondary color.