Improve Tap Handling
Sometimes the tap handling wasn't applied to the correct component in a
list, e.g. in nearly all list items we used a content padding so that
taps on the padding were not detected. This is now fixed, so that taps
on list items are handling idependent of where the user taps on the
item.