Allow Sponsoring via In-App Purchases (#453)
It is now possible to sponsor the development of kubenav via In-App
Purchases on iOS and Android. When a user isn't a sponsor, we display a
small sponsor icon in the actions menu on the home screen and a sponsor
section in the settings. We also display a sponsor banner in the
settings, which can be removed by a user and will only be displayed
after 7 days.

If a user is a sponsor, all these references are removed.

Note: There are no features behind this sponsoring, besides that all the
refrences to becomme a sponsor are removed from the app. I also do not
plan to put anything behind a stupid paywall. This is just another way
to support the development of the app. Besides that I also just wanted
to try In-App Purchase to learn sth. new and hope that I found a good
way to not disturb anyone with this change. If you feel annoyed by it
please let me know, so that we can adjust the feature to become less
annoying for everyone. Thanks for your understanding.