Support Port-Forwarding via Services
It is now possible to establish a port-forwarding session via a service.
To do this a user can now click on a port in the service details view.
We will then use the service selector to get the corresponding pods and
will establish a port forwarding session to the first port.