Make Sponsor Banner Less Annoying
This commit fixes a bug with the sponsor banner in the settings screen.
When the user clicks on the sponsor banner and wants to hide it via the
"Not Now" button, it shouldn't be displayed for the next seven days, but
I forgot to store this value in the settings, so that the sponsor banner
was displayed with every app restart. This should now be fixed.

I also added a "Restore Purchases" button, which is required for the
Apple AppStore. When the button is clicked we restore all previous
purchases of the user, similar to how it is handled during the app
start.