Fix Show all Pods Running on Node
In the node details view it is possible to view all pods which are
running on the currently selected node. There was a bug which filtered
the pods shown after clicking on "View all" by the currently selected
namespace for the cluster. This should now be fixed by checking if the
selector is used to select all pods for a node. If this is the case we
are ignoring the current cluster namespace to get the pods.