Get Logs for Deployment, DaemonSet and StatefulSet (#470)
It is now possible to view the logs for a Deployment, DaemonSet and
StatefulSet. For this the details view of these resources contains a new
logs action, which will open a modal, where the user can select the Pods
for which he wants to view the logs. Once the user made a selection the
already known logs modal is opened where the user can specify the other
settings to get the logs (e.g. container name).