Disable Navigation to Current Page
Until now it was possible to navigate to the same page via the bottom
navigation bar. This is not possible anymore. This was removed, because
it was a bit confusing for users, e.g. a user was on the home page and
clicked the home icon, the navigation animation was displayed, but the
user was still on the same page.

This doesn't effect the behaviour that the bottom navigation bar can
also be used to go back to the root page, e.g. a user views the details
of a Kubernetes resource, he can still go back to the resources list by
clicking on the resources icon in the navigation bar.