Add Full Height Modal Setting (#474)
This commit adds a new "fullHeightModals" setting in the app settings,
which allows a user to enabled full height modals (modals which takes
the entire screen size and not just 75%). This setting isn't enabled by
default, because it also comes with the drawback that the modal header
is moved behind the safe area.

Hopefully this should be fixed by the "safeArea" property for modal
bottom sheets in Flutter 3.7.0 (currently we are not able to update
because some dependencies are not working with the new version).