Add Actions for Resources in List View
It is now possible to view the actions like gettings logs for a Pod,
scaling a Deployment, etc. in the list view of the resources. The
actions can be opened via a double tap on the resource. This allows
users to use actions without going to the details page of a resource.