Re-add Rancher Support (#481)
Re-add support for the Rancher Cloud Provider, which was already
available in kubenav v3. The Rancher provider requires the address of
the Rancher API, a username and password. A user can also provide a
token to access the Rancher API or he can generate a token via the sign
in button.

The optional "Allow Insecure Connections" option can be used when
Rancher runs with a self signed certificate.