Add Support for Streaming Logs
It is now possible to listen on a stream of logs for a Pod / Container.
For this we had added a new route to the http server, which can be used
to create a WebSocket connection to stream the logs.

We have keept the original logs implementation, because when the user
just wants to see the logs without following them we can still use this
implementation, which has the advantage that we do not have to start a
http server.