Improve Continuous Delivery GHA for macOS
The continuous delivery GitHub Action requires some secrets for the
macOS build, so that the action fails for PRs from external
contributors. This PR adds a check to all steps, which require some
secrets to omit them for PRs where the secrets are not present, so that
the action will be green.