Automatic Select Theme based on OS Settings (#492)
The theme is now automatically selected based on the users OS settings.
The new logic is applied when a user selects the "auto" option in the
app settings. For users which already selected the "light" or "dark"
theme the logic is not applied and they have to opt in for the "auto"
option.