Add "Classic Mode"
The new "classic mode" implements the design of version 3 of kubenav,
which is prefered by some users. Since the required changes and the
maintance effort shouldn't be that high we can support the new design
and the old one as opt in. The changes which are applied, when a user
opt in for the "classic mode" are the followings:

- Remove the bottom navigation bar and add a sidebar for the navigation
  in the app
- Remove the resource actions header (the one were a user can view the
  logs of a container or scale a deployment) and move the actions to a
  menu, which can be accessed via the app bar.

See #477