Allow Custom Redirect URL for OIDC and Google (#495)
It is now possible to specify a custom redirect url for the OIDC and
Google provider. Since we only have to get the "code" parameter to
finish the authentication flow, it doesn't matter if the site is hosted
at kubenav.io or a user hosts them by themself. For the user it has not
to allow kubenav.io as redirect url.

The following html files can be hosted at the provided redirect url to
make it easier to copy the token:

- https://github.com/kubenav/kubenav/blob/main/kubenav.io/auth/oidc.html
- https://github.com/kubenav/kubenav/blob/main/kubenav.io/auth/google.html