Fix Y-Axis Labels in Live Metrics
The y axis in the live metrics wasn't formatted that well, because we
always rounded to full numbers the displayed labels were duplicated.
This should now be fixed, because we are checking the value of the label
before it is rendered. This means that we add 2 decimal places when the
value of the label is very large (full cores for CPU or Gi for Memory).