Add Support for Multiple AWS SSO Accounts
It is now possible to add multiple AWS SSO Accounts and Roles at once,
by only specifing the AWS SSO start url and region. We then
automatically load all the accounts and the roles, so that a user can
select which accounts and roles should be added to the app. Each
selected account/role will be added as a new AWS SSO provider, so that
we can use the already existing flow to add the clusters and to handle
the authentication against the Kubernetes API of the added clusters.