Fix "Classic Mode" when User doesn't have any Clusters (#525)
When a user doesn't have any clusters configured and selects one of the
resources from the sidebar in the "classic mode" the app crashes. This
was caused, since we assumed that a user must have an active cluster
when visiting the page, which is only true for the default layout but
not for the "classic mode".

This is now fixed by adding conditional checks where required in the
resources list widget and by throwing an error before making an API
request when the list of clusters is empty.