Fix Restart of Deployments (#542)
Under some circumstances it could happen that the restart action for
Deployments, StatefulSets and DaemonSets deleted all existing
annotations from "spec.template.metadata.annotations". This should now
be fixed by checking if there are already other annotations besides the
annotation we have to add. If this is the case we now apply another JSON
patch to add our annotation.

Fixes #540