Update Flutter to Version 3.16.0 (#555)
Update Flutter to version 3.16.0 which is the latest release and update
all Flutter packages to their latest version. This also includes a new
version of "flutter_lints". The errors mentioned by the new linter
version were also fixed within this PR.

With the update some styling parameters were changed, so that the theme
must be adjusted. This will be done in a follow up PR to not add more
changes.