Update Flutter to Version 3.19.5
Upgrade the used version of Flutter to 3.19.1 and all used packages to
their latest version.

For the Android Version of the app, we also had to adjust some files,
because of the "Deprecated imperative apply of Flutter's Gradle
plugins". The steps we had to run for this can be found at
https://docs.flutter.dev/release/breaking-changes/flutter-gradle-plugin-apply