Use Background Threads on iOS (#584)
For the Kubernetes API calls and for the different cloud providers we
are using Go which is called via platform channels in Flutter. In the
Android implementation we already executed the called native handlers in
background threads and this commit adds the same functionality to our
iOS implementation. This wasn't done earlier, because we run in the
issue mentioned in https://github.com/flutter/flutter/issues/118832,
which was fixed and included in Flutter v3.19.0.