Improve Theme Handling
Instead of using the `ThemeRepository` to define and return the correct
theme for dark / light mode, we are now defining a light and dark theme,
with theme extensions. This allows us to use `Theme.of(context)` to use
the color of the currently selected mode within the app.

The `ThemeRepository` is now only used to get the theme which is
currently selected by the user.