Remove `onPlatformBrightnessChanged` Function
In the `AppMaterialApp` widget we were using the
`onPlatformBrightnessChanged` function to listen for brightness changes
in the app, so we could update the theme when the user has selected the
`auto` theme option.

After the rework of the theme handling in #597, this is not needed
anymore, because when the `auto` option is selected, we set the theme
mode to `ThemeMode.system` in the `MaterialApp` widget so that the theme
is automatically selected on brightness changes.