Improve Terminal Handling
Instead of using tabs to show multiple terminals within one view, we are
now using a similar logic as for the port forwarding. This means that a
user can select the terminal he wants to view via the floating action
button, after that we show the list of terminals as actions, were the
user can select the terminal. Once a terminal was select the user can
open or close the terminal. When the terminal is opened we only show the
selected terminal.