Helm: Add Support for Release Rollbacks (#609)
This commit adds support to rollback Helm releases. When a user selects
the rollback option he can set the same flags as via the Helm cli.

This is possible, because we reworked the Helm plugin, to use the Helm
Go package for all actions now. Within this rework we also adjusted the
naming of the existing Helm function to be more clear on what they are
doing.