Add Home Page Configuration (#618)
The home page of the app can now be configured by the user. It is
possible to enable and disable the metrics and warnings component and it
is possible to select the scope for the warnings component. This means
that when a user doesn't have the permissions to show the warnings from
all namespaces he can select the "Use Selected Namespace" options to
only show the warnings from his currently selected namespace.

This also allows us to add more components to the home page in the
future, e.g. to show unhealthy workloads, which can then be enabled /
disabled by the user.