Adjust `KubernetesService().getRequest()` Function
Adjust the `KubernetesService().getRequest()` function, to return a
`String` instead of a `Map<String, dynamic>`. Because of this adjustment
we have to run `json.decode` always afer we are using the function, but
this will allow us to run the decode and parsing function in it's own
Isolate in the future to improve the perfromance.