Flux Plugin: Improve Performance
This commit improves the performance of the Flutter plugin, by using the
`compute` function to decode the json responses from the Kubernetes API.

For the implementation we had to rework most of the parts of the plugin,
so that we can decode the returned response in the correct Flux
resource.

Because of this rework we were also not able to reuse all the default
widgets from the Kubernetes resources implementation, so that we had to
add the following widgets:

- `PluginFluxShowYaml`: To show the manifest of the Flux resource. In
  opposite to the existing widget, we pass in a `encodeItem` function,
  which is executed inside of the `compute` function to improve the
  render performance.
- `DetailsItemMetadata`: This new widget uses the
  `IoK8sApimachineryPkgApisMetaV1ObjectMeta` model instead of a
  `Map<String, dynmic>` type.
- `DetailsItemConditions`: This widget uses the
  `IoK8sApimachineryPkgApisMetaV1Condition` model instead of a
  `Map<String, dynamic>` type. For this we also had to adjust the
  conditions type in all Flux resources.

This widgets are intended to be reused later, when we will rework the
Kubernetes resources pages, to also improve the rendering performance
there.

Last but not least we also adjusted the `DetailsItemWidget` widget to
ignore `null` values and empty lists.