Rework Resources Handling
Similar to the changes made for the Flux plugin in #631, we also
adjusted the resources handling for all Kubernetes resources. This
should improve the performance while rendering the different pages and
it also makes working with the resources easier since we are now always
knowing the exact type of the resource.