Add Support for CRDs in `goToReference` Function
We are now supporting CRDs in the `goToReference` function, this means
if a event or the ownership of a resource is related to a CRD, the user
can now directly go to the CRD, similar to how it was already possible
for the standard Kubernetes resources.

Since we have to get all CRDs from the Kubernetes API server for this,
we implemented a `CRDsCacheRepository` to cache the CRDs, so that we
do not have to fetch the CRDs everytime, a user wants to open a
reference.