Fix Unnecessary Rerenderings when Using Bookmark Action (#642)
When a user bookmarked a list of resources or a single resource in the
details view, the whole page was rerendered, because we have to listen
for updates on the `BookmarksRepository`. To avoid that the whole view
is rerendered we moved the actions to a seperate widget and only listen
there for changes of the `BookmarksRepository`.

We also changed the `additionalActions` property, to not except a list
of actions, but only a `refresh` function, which can be defined in the
details view and which can be `null` in the list view.