Show Workload Status on Home Page
It is now possible to show the status for Pods, Deployments,
StatefulSets, DaemonSets and Jobs on the home page. This feature can be
enabled seperatly for each workload in the home page settings.

When enabled we display for each workload the total number of resources
and the count for each resource status. When a user click on the
workload, he can directly view all the resources with the selected
status.

For this the `ResourcesList` widget now contains a `status` parameter,
which is set to the user selected status to filter the resources
accordingly.