Improve Events Handling on Home Page
This commit improves the soring of events on the home page. We are now
using the `lastTimestamp`, `eventTime` or `creationTimestamp` for
sorting the returned events.

The list and preview items for events now also contain the Namespace,
where the event is comming from and for the "Last Seen" field we are
using the fields mentioned above.