Add `CertificateSigningRequests` Resource
This commit adds support for the `CertificateSigningRequest` resource.
This means that users can view CSRs within the app. It is also possible
to approve or deny a CSR within the app.

We also added a new `DateTimeExtension`, to format the time in a format
required by the Kubernetes API server. For this we also moved the
`_rFC3339Nano` function required by the Flux plugin to the newly added
extension.