Use Long Press Instead of Double Tap for Actions
The secondary actions for various items in the app can now be triggered
via a long press instead of a double tap. We decided to go with the long
presse, because most of the other apps are also using a long press
instead of a double tap.

When the user long presses a item, the user will now also get some
haptic feedback, to make sure that a secondary action is supported by
the item.