Fix Modal Bottom Sheet Width for Terminal (#653)
For terminals and logs the modal bottom sheet should always use the full
width of the screen, but on larger screens (e.g. iPad) it wasn't using
the full screen width.

This is now fixed by adding a new `fullScreen` parameter to the
`showModal` function, which is `false` by default, but can be set to
`true` for the terminal and logs, so that they are using the full screen
width and height.