Add SSH Action for Nodes
The new SSH action for Nodes can be used to create a privileged Pod on
the selected node, with similar rights as a user would connect via SSH
to the node.

The action will create the Pod on the selected Node, afterwards a user
can view the Pod and can create a new terminal, to view all the
processes and files running on the Node.