Improve Spacing for Horizontal Lists
The spacing was a bit off compared with all other widgets in the
horizontal lists, like we are using them for the workload conditions,
cluster view in the settings and the workloads overview.

This is now fixed, so that the lists behave similar to the other
horizontal lists, with the same spacing.