Add Batch Delete Action (#663)
It is now possible to delete multiple resources at once. For that we
added a new delete action to the list view, which will show the user a
list of all the resources. A user can then select the items which should
be deleted and we will run the delete call for all the selected items.