Add Swipe Action for Namespaces
It is now possible to show the delete action for Namespaces in the
settings by swiping the namespace to the left. This should make it
easier fo users to "find" the secondary action for namespaces and should
also adopted for other components where the secondary actions could be
triggered via a long press.

For these new gestures we added the `flutter_slidable` package, which
makes it easier to implement this feature. We also changed the `onError`
color for the light and dark theme to match the `onPrimary` color.