Add Swipe Action for Bookmarks
The action to delete a bookmark, can now be triggered by swiping a
bookmark item to the left. This is implemented as an alerternative to
the long press to delete a bookmark and should make it easier for users
to find the action.

This commit also changes the function to delete a bookmark to an async
function so that we can await the deletion of a bookmark.