Improve Design of `Slidable` Widget
The `Slidable` widget has rounded corners now and the background of the
item which is swiped is now in the same color as the first action.

We also moved the `Slidable` wiedget to the `AppListItem` widget, which
now has a new property `slidableActions` which can be used to define the
actions, which are triggered via the swipe gesture.