Add Pod Eviction Action (#684)
It is now possible to evict Pods via the Kubernetes eviction API within
the app. For this the Pod details view contains a new `Evict` action,
which can be used to evict a Pod.

This was added as alternative to the delete action, because the eviction
API also respects a possible defined PDB, which might be useful in some
sitiuations.