Fix Creation of Jobs from CronJobs (#688)
It could happen that the creation of Jobs from a CronJob failed via the
corresponding action, because some fields which can not be present in
the Job spec, where null or an empty list. This is now fixed by removing
these values before making the Kubernetes API call.

Besides that we are now using the same logic as kubectl to create the
Job, see https://github.com/kubernetes/kubectl/blob/d3ad75324522280fa8bccd7ad949b34336f5fc84/pkg/cmd/create/create_job.go#L254

For that we also added a new text field, which can be used to specify
the name of the job which should be created. To make it easier for
users, the field is prefilled with the CronJob name and the `-manual`
suffix.