Add Actions to Suspend / Resume CronJobs
It is now possible to suspend and resume CronJobs via the newly added
actions. This should make it easier for users to suspend / resume
CronJobs, because they do not have to edit the Yaml files anymore.

For the actions we create a JSON patch were we set the
`cronJob.spec.suspend` field to `true` or `false`, depending on the
selected action. This is very similar to the suspend / resume action in
the Flux plugin.

We also change the icon used for the "Create Job" action, so that we can
reuse the icons from the Flux suspend / resume actions.