Use `scale` Instead of `patch` for Scaling Resources
This commit changes the behaviour of the scale action, which can be used
to scale a Deployment, StatefulSet or ReplicaSet. Until now we made a
`patch` request to the Kubernetes API to adjust the number of replicas.
Now we are using the `/scale` endpoint to achieve the same.

This was done, so that only scaling can be allowed in the RBAC rules for
a user, see #701.