Add State Parameter to OIDC Login URL (#710)
Add the state parameter to the OIDC login url. We show a random
generated string in the OIDC login url, which is shown in the provider
config and on the redirect page. Normally this should be veryfied
automatically, but since we do not redirect to the app a user should
manually compare the state parameters from the provider configuration
and redirect page.