Fix OIDC Provider for Google (#722)
When a user wants to use Google as OIDC provider it wasn't working,
because we automatically add the `offline_access`, which isn't supported
by Google.

This is now fixed, by checking if the provided "Discovery URL" is
`https://accounts.google.com`. When this is the case we do not add the
`offline_access` scope.