Simplify Re-Authentication for OIDC Provider (#728)
This commit simplifys the re-authentication within the OIDC provider.
This is required, because it looks like not all providers are supporting
refresh tokens, so that once the access token is expired a user must
re-authenticate.

When we detect this situation (access token expired + no refresh token)
we throw a custom error message. Based on this message we then show a
custom error widget, where the user can sign in again and after
providing the new `code` parameter we can get a new access token.