OIDC: Add Option to Disable Default Scopes (#769)
The OIDC provider now contains an options to disable the default scopes
`openid` and `offline_access`. This was added because the
`offline_access` scope is not supported by some OIDC providers. E.g.
Google doesn't support it and we worked arround this in the past by
checking if the provider url was related to Google. This solutions does
not work for urls where the service is self hosted with a custom url
like mentioned in #750. Therefore it is now possible to skip the default
scopes at all.

We can not simply remove the default scopes at all, because it could
break existing configuration, so that the explicit option was added and
the default value is that the default scopes are added.

Fixes #750