Catch Panics when Creating a New Kubernetes Client (#780)
In the past we had reports that the app crashes, when a invalid
Kubeconfig is used within the app. In the past this was caused by using
a token, username and password in the configuration. Since this error is
already catched, we now catch all panics in the `NewClient` function and
return the panic message as error.

If we still see crashes of the app afterwards, we might have to do this
for all Go functions.