Improve Pod Status for Restarts (#781)
Until now we always showed a warning for Pods where the restart count
was larger then zero. This is now improved, so that we also include the
last time when the Pod was restarted. This means if a Pod has a restart
count greater than 0 and the last restart was within the last 24 hours
we show a warning. If the restart count is greater than 0 and the last
restart was not within the last 24 hours we do not show a warning.