Catch Panics in Go Functions (#784)
This is a follow up PR for #780, which didn"t resolved #776. Because it
is still not clear where the app crashes exactly when a broken
Kubeconfig is used, we now catch the panics in all called Go functions
and check if the result is empty in the Dart code to return an "Unknown
Error" error message.

If this still doesn't solve the issue, the error might be within the
Dart code. I have no idea where an error can happen there which crashes
the app.

We can also try to provide a entrypoint for the app which can be used to
delete the whole cluster and provider configuration.