Add Reset Action
Add a reset action, which can be called via quick actions (long press on
the app icon). This action is used to open the `Reset` widget instead of
the normally shown widgets after the app is started. The `Reset` widget
then provides an option to delete all existing providers, clusters and
settings.

This is a workaround for when the app crashs, because something is wrong
with a configured cluster. This is a hacky workaround, but should be
better then that users are not able to use the app at all.