Improve OIDC Redirect Page
When the OIDC flow fails, we still displayed the button to copy the code
also when there was no code parameter avaiable. This is now changed to
only show the button when there is a code parameter.

If there is no code parameter but an error parameter, we now will also
show the error.