Improve Container Handling for Logs and Terminal
This commit improves the container handling in the logs view, to show
the containers before the init containers as it was done in #802 for the
logs view.

Besides that this commit also adds the emphermal containers to the list
of containers which can be selected in the logs view. This was done
because they can also be selected for in the terminal view.